Split out ThreePartNamesParser
Splits out ThreePartNamesParser helper class in a separate file, renames
it to ThreePartName, and changes the regex for spliting fully qualified
procedure names (ex: database.owner.object) to be compiled once and
stored in a static variable rather than rather than repeatedly on each
invocation.

As it is not part of the public API, the ThreePartName class is not
marked public.