Refactor SQLServerParameterMetaData to use ThreePartName
Refactor SQLServerParameterMetaData to use ThreePartName for parsing and
splitting procedure identifiers into separate components rather than
implement its own parsing logic.

Renames the method to parseProcIdentifier(...) and changes the return
to not include the trailing comma so that it can be reused elsewhere
without assuming there is a trailing parameter.