Fix | Replace timeout handling to use SharedTimer (#920)
* Replace timeout handling to use SharedTimer

Replaces the timeout handling for basic and bulk TDS commands to use a new SharedTimer
class. SharedTimer provides a static method for fetching an existing static object or
creating one on demand. Usage is tracked through reference counting and callers are
required to call removeRef() when they will no longer be using the SharedTimer. If the
SharedTimer does not have any more references then its internal ScheduledThreadPoolExecutor
will be shutdown.

The SharedTimer is cached at the Connection level so that repeated invocations do not
create new timers. Connections only create timers on first use so if no actions involve
a timeout then no timer is fetched or created. If a Connection does create a timer then
it will be released when the Connection closed.

Properly written JDBC applications that always close their Connection objects when
they are finished using them should not have any extra threads running after they are
all closed. Applications that do not use query timeouts will not have any extra threads
created as they are only done on demand.  Applications that use timeouts and use
a JDBC connection pool will have a single shared object across all JDBC connections as
long as there are some open connections in the pool with timeouts enabled.

Interrupt actions to handle a timeout are executed in their own thread. A handler thread is
created when the timeout occurs with the thread name matching the connection id of the
client connection that created the timeout. If the timeout is canceled prior to the interrupt
action being executed, say because the command finished, then no handler thread is created.

Note that the sharing of the timers happens across all Connections, not just Connections
with the same JDBC URL and properties.

* Normalize newlines in TimeoutTest

* Move TimeoutTest to parent package so that it can access package private members

* Centralize SharedTimer thread checks in TimeoutTest

Centralize checks for SharedTimer thread state in TimeoutTest to happen before and after
all test invocations in TimeoutTest via JUnit @Before and @After annotations. Each test
now requires that the SharedTimer thread not be running before the test starts and after
the test completes.

Also expands the thread name prefix in SharedTimer to package private so that TimeoutTest
can reference it rather than having its own copy fo the string.

* Centralize and reduce timeout value used in TimeoutTest

Centralizes constant timeout value used in TimeoutTest and reduces it
from ten seconds down to two seconds to speed up tests.

* Centralize SQL for WAIT FOR DELAY in TimeoutTest

Centralizes the SQL command for WAIT FOR DELAY in TimeoutTest and changes
the format of the delay to use hour:minute:second for clarity.

* Clean up TimeoutTest to use assertThrows(...)

Changes test in TimeoutTest to use assertThrows(...) rather than manually
checking for the thrown exception.

Also cleans up and adds internal helpers for creating a connection and changes
internal runQuery(...) to be void rather than returning a boolean.

* Add additional SharedTimer tests

Adds additional tests to verify creation and shutdown of SharedTimer core thread
when user commands specify a query timeout.

* Improve waiting for SharedTimer thread stoppage

Change waiting for SharedTimer thread to stop to happen in a loop with periodic
checks to see if the thread has stopped up to a max amount of waiting (10 seconds).

* Remove System.err.println(...) in TdsTimeoutTask

* Rename static SharedTimer singleton to instance

* Replace SharedTimer ThreadFactory with inline lambda

* Add comments for SharedTimer

* Rename TdsTimeoutTask to TDSTimeoutTask