feat: Add custom SocketFactory
Adds a connection property, socketFactoryClass, to override the SocketFactory
used to create Sockets for driver connections. If unspecified, the default, then
the system default SocketFactory.getDefault() will be used. If a class name is
specified then it must be assignable to javax.net.SocketFactory.

An optional argument can be specified via the property socketFactoryConstructorArg.
If specified then the single String argument constructor of the class will be
invoked. Otherwise the no-argument constructor will be invoked.