Added connection properties to specify custom SocketFactory (#1217)
* misc: Add Util.newInstance(...) helper

Adds a Util.newInstance(...) helper to instantiate a class by name with
an optional argument and replaces similar code to create a custom
TrustManager with the new helper.

* feat: Add custom SocketFactory

Adds a connection property, socketFactoryClass, to override the SocketFactory
used to create Sockets for driver connections. If unspecified, the default, then
the system default SocketFactory.getDefault() will be used. If a class name is
specified then it must be assignable to javax.net.SocketFactory.

An optional argument can be specified via the property socketFactoryConstructorArg.
If specified then the single String argument constructor of the class will be
invoked. Otherwise the no-argument constructor will be invoked.

* fix: Correct typo in exception message

* test: Add a test for an invalid socketFactoryClass

* error messages

* error message format

* Improve error message and format files

* Update error message

Co-authored-by: v-reye <v-reye@microsoft.com>
Co-authored-by: ulvii <v-ulibra@microsoft.com>
Co-authored-by: Peter Bae <v-hyba@microsoft.com>